+++ database/migrations/2025_05_21_102302_create_commande_medictations_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('commande_id')->constrained()->cascadeOnDelete();
             $table->foreignId('medication_id')->constrained()->cascadeOnDelete();
-            $table->decimal("total_price");
+            $table->integer("total_price");
             $table->integer("quantity");
             $table->timestamps();
         });
